Can fix the number of float data

+++ UD Unggas Sari Utama/UD Unggas Sari Utama/Model/PemesananModel/DetaiPemesananModel.swift
@@ -12,7 +12,7 @@ struct DetailPemesananModel: Codable, Hashable{
     var id: String
     var pemesanan_id: String
     var produk_id: String
-    var jumlah_produk: Int
+    var jumlah_produk: Float
     var harga_produk: Int
     var sub_harga: Int
     
@@ -31,7 +31,7 @@ struct DetailPemesananModel: Codable, Hashable{
         pemesanan_id = try values.decode(String.self, forKey: .pemesanan_id)
         produk_id = try values.decode(String.self, forKey: .produk_id)
         let jumlahString = try values.decode(String.self, forKey: .jumlah_produk)
-        jumlah_produk = Int(jumlahString) ?? 0
+        jumlah_produk = Float(jumlahString) ?? 0
         let hargaString = try values.decode(String.self, forKey: .harga_produk)
         harga_produk = Int(hargaString) ?? 0
         let totalString = try values.decode(String.self, forKey: .sub_harga)
